requirements.txt
# Core dependencies
openai>=1.12.0
tweepy>=4.14.0
feedparser>=6.0.10
python-dotenv>=1.0.0
requests>=2.31.0
tenacity>=8.2.3
youtube-transcript-api>=0.6.2
rich>=13.0.0

# Viral content machine dependencies
sqlalchemy>=2.0.0
flask>=2.3.0
pillow>=10.0.0
matplotlib>=3.7.0
textstat>=0.7.3
nltk>=3.8.1
transformers>=4.30.0
librosa>=0.10.1
schedule>=1.2.0

# Autonomous operation dependencies
psutil>=5.9.0

# Sprint 2: Multi-model AI and quality improvements
anthropic>=0.18.0
google-generativeai>=0.3.0

# Optional: Local Whisper for unlimited transcription
openai-whisper>=20231117

# OSINT Handle Finder dependencies
beautifulsoup4>=4.12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0